fix signatureMessageData error message
fix for error: cannot allocate an array of constant size 0
was unsigned long long
now uint64_t  
http://stackoverflow.com/questions/18971732/difference-between-long-long-long-long-int-and-long-long-int-in-c

+++ src/ed25519.cc
@@ -74,7 +74,7 @@ NAN_METHOD(Sign) {
 	Handle<Object> message = args[0]->ToObject();
 	const unsigned char* messageData = (unsigned char*)Buffer::Data(message);
 	size_t messageLen = Buffer::Length(message);
-	unsigned long long sigLen = 64 + messageLen;
+	uint64_t sigLen = 64 + messageLen;
 	unsigned char signatureMessageData[sigLen];
 	crypto_sign(signatureMessageData, &sigLen, messageData, messageLen, privateKey);
